+++ src/Model/AbstractNamedDomIdentifier.php
@@ -1,0 +1,33 @@
+<?php declare(strict_types=1);
+
+namespace webignition\BasilTranspiler\Model;
+
+use webignition\BasilCompilationSource\VariablePlaceholder;
+use webignition\BasilModel\Identifier\DomIdentifierInterface;
+
+abstract class AbstractNamedDomIdentifier implements NamedDomIdentifierInterface
+{
+    private $identifier;
+    private $placeholder;
+
+    public function __construct(DomIdentifierInterface $identifier, VariablePlaceholder $placeholder)
+    {
+        $this->identifier = $identifier;
+        $this->placeholder = $placeholder;
+    }
+
+    public function getIdentifier(): DomIdentifierInterface
+    {
+        return $this->identifier;
+    }
+
+    public function getPlaceholder(): VariablePlaceholder
+    {
+        return $this->placeholder;
+    }
+
+    public function includeValue(): bool
+    {
+        return null !== $this->identifier->getAttributeName();
+    }
+}
